refactor(loon-core): split commit internals into focused modules
