+++ crates/loon-core/src/commit/request.rs
@@ -1,0 +1,18 @@
+use super::{CommitOp, Precondition};
+use loon_api::v0::CommitAnnotations;
+use loon_api::{CommitId, FenceToken, NamespaceId};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct CommitRequest {
+    pub namespace_id: NamespaceId,
+    pub commit_id: CommitId,
+    pub writer_id: String,
+    pub writer_fence_token: FenceToken,
+    pub ops: Vec<CommitOp>,
+    pub preconditions: Vec<Precondition>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub annotations: Option<CommitAnnotations>,
+}
